api/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.6
python-dotenv==1.0.0
SQLAlchemy==2.0.19
PyMySQL==1.1.0
